+++ api/argsBuilder.go
@@ -4,19 +4,6 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-// TODO: REMOVE
-func streamFieldsArgs(ctx *gin.Context) (flowUUID string, streamUUID string, producerUUID string, consumerUUID string, writerUUID string) {
-	var args Args
-	var aType = ArgsType
-	var aDefault = ArgsDefault
-	args.FlowUUID = ctx.DefaultQuery(aType.FlowUUID, aDefault.FlowUUID)
-	//args.StreamUUID = ctx.DefaultQuery(aType.StreamUUID, aDefault.StreamUUID)
-	args.ProducerUUID = ctx.DefaultQuery(aType.ProducerUUID, aDefault.ProducerUUID)
-	args.ConsumerUUID = ctx.DefaultQuery(aType.ConsumerUUID, aDefault.ConsumerUUID)
-	args.WriterUUID = ctx.DefaultQuery(aType.WriterUUID, aDefault.WriterUUID)
-	return args.FlowUUID, "todo remove", args.ProducerUUID, args.ConsumerUUID, args.WriterUUID
-}
-
 //withFieldsArgs
 func withFieldsArgs(ctx *gin.Context) (field string, value string) {
 	var args Args
@@ -45,22 +32,6 @@ func parentArgs(ctx *gin.Context) (AddToParent string) {
 	var aDefault = ArgsDefault
 	args.AddToParent = ctx.DefaultQuery(aType.AddToParent, aDefault.AddToParent)
 	return args.AddToParent
-}
-
-//withConsumerArgs
-func withConsumerArgs(ctx *gin.Context) (askResponse bool, askRefresh bool, write bool, updateProducer bool) {
-	var args Args
-	var aType = ArgsType
-	var aDefault = ArgsDefault
-	args.AskRefresh = ctx.DefaultQuery(aType.AskRefresh, aDefault.AskRefresh)
-	args.AskResponse = ctx.DefaultQuery(aType.AskResponse, aDefault.AskResponse)
-	args.Write = ctx.DefaultQuery(aType.Write, aDefault.Write)
-	args.UpdateProducer = ctx.DefaultQuery(aType.UpdateProducer, aDefault.UpdateProducer)
-	askRefresh, _ = toBool(args.AskRefresh)
-	askResponse, _ = toBool(args.AskResponse)
-	write, _ = toBool(args.Write)
-	updateProducer, _ = toBool(args.UpdateProducer)
-	return askRefresh, askResponse, write, updateProducer
 }
 
 func buildFlowNetworkArgs(ctx *gin.Context) Args {
@@ -154,6 +125,12 @@ func buildProducerArgs(ctx *gin.Context) Args {
 	if value, ok := ctx.GetQuery(aType.StreamUUID); ok {
 		args.StreamUUID = &value
 	}
+	if value, ok := ctx.GetQuery(aType.Name); ok {
+		args.Name = &value
+	}
+	if value, ok := ctx.GetQuery(aType.ProducerThingUUID); ok {
+		args.ProducerThingUUID = &value
+	}
 	return args
 }
 
@@ -171,6 +148,7 @@ func buildDeviceArgs(ctx *gin.Context) Args {
 	var args Args
 	var aType = ArgsType
 	var aDefault = ArgsDefault
+	args.WithPriority, _ = toBool(ctx.DefaultQuery(aType.WithPriority, aDefault.WithPriority))
 	args.WithPoints, _ = toBool(ctx.DefaultQuery(aType.WithPoints, aDefault.WithPoints))
 	args.WithTags, _ = toBool(ctx.DefaultQuery(aType.WithTags, aDefault.WithTags))
 	return args
@@ -188,8 +166,14 @@ func buildPointArgs(ctx *gin.Context) Args {
 func buildWriterArgs(ctx *gin.Context) Args {
 	var args Args
 	var aType = ArgsType
+	if value, ok := ctx.GetQuery(aType.ProducerUUID); ok {
+		args.ProducerUUID = &value
+	}
 	if value, ok := ctx.GetQuery(aType.WriterThingClass); ok {
 		args.WriterThingClass = &value
+	}
+	if value, ok := ctx.GetQuery(aType.SourceUUID); ok {
+		args.SourceUUID = &value
 	}
 	return args
 }