+++ api/syncWriter.go
@@ -6,7 +6,9 @@ import (
 )
 
 type SyncWriterDatabase interface {
-	SyncWriter(fn *model.SyncWriter) (*model.WriterClone, error)
+	SyncWriter(body *model.SyncWriter) (*model.WriterClone, error)
+	SyncCOV(body *model.SyncCOV) error
+	SyncWriterAction(body *model.SyncWriterAction) error
 }
 
 type SyncWriterAPI struct {
@@ -17,4 +19,16 @@ func (a *SyncWriterAPI) SyncWriter(ctx *gin.Context) {
 	body, _ := getBodySyncWriter(ctx)
 	q, err := a.DB.SyncWriter(body)
 	responseHandler(q, err, ctx)
+}
+
+func (a *SyncWriterAPI) SyncCOV(ctx *gin.Context) {
+	body, _ := getBodySyncCOV(ctx)
+	err := a.DB.SyncCOV(body)
+	responseHandler(nil, err, ctx)
+}
+
+func (a *SyncWriterAPI) SyncWriterAction(ctx *gin.Context) {
+	body, _ := getBodySyncWriterAction(ctx)
+	err := a.DB.SyncWriterAction(body)
+	responseHandler(nil, err, ctx)
 }