+++ database/producer.go
@@ -3,14 +3,11 @@ package database
 import (
 	"encoding/json"
 	"errors"
-	"fmt"
 	"github.com/NubeIO/flow-framework/api"
-	"github.com/NubeIO/flow-framework/eventbus"
 	"github.com/NubeIO/flow-framework/model"
+	"github.com/NubeIO/flow-framework/src/client"
 	"github.com/NubeIO/flow-framework/utils"
 	log "github.com/sirupsen/logrus"
-	"gorm.io/datatypes"
-	"time"
 )
 
 type Producer struct {
@@ -31,6 +28,15 @@ func (d *GormDatabase) GetProducer(uuid string, args api.Args) (*model.Producer,
 	query := d.buildProducerQuery(args)
 	if err := query.Where("uuid = ?", uuid).First(&producerModel).Error; err != nil {
 		return nil, err
+	}
+	return producerModel, nil
+}
+
+func (d *GormDatabase) GetOneProducerByArgs(args api.Args) (*model.Producer, error) {
+	var producerModel *model.Producer
+	query := d.buildProducerQuery(args)
+	if err := query.First(&producerModel).Error; err != nil {
+		return nil, query.Error
 	}
 	return producerModel, nil
 }
@@ -122,86 +128,92 @@ type Point struct {
 	Priority model.Priority `json:"priority"`
 }
 
-// ProducerWriteHist  update it
-func (d *GormDatabase) ProducerWriteHist(uuid string, writeData datatypes.JSON) (*model.ProducerHistory, error) {
+func (d *GormDatabase) ProducersPointWrite(point *model.Point) error {
+	producerModelBody := new(model.Producer)
+	producerModelBody.CurrentWriterUUID = point.UUID // TODO: check current_writer_uuid is needed or not
+	producers, _ := d.GetProducers(api.Args{ProducerThingUUID: &point.UUID})
+	for _, producer := range producers {
+		err := d.producerPointWrite(producer.UUID, point, producerModelBody)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (d *GormDatabase) producerPointWrite(uuid string, point *model.Point, producerModelBody *model.Producer) error {
+	producerModel, err := d.UpdateProducer(uuid, producerModelBody) // TODO: check current_writer_uuid is needed or not
+	if err != nil {
+		log.Errorf("producer: issue on update producer err: %v\n", err)
+		return errors.New("issue on update producer")
+	}
+
+	syncCOV := model.SyncCOV{Priority: point.Priority}
+	err = d.TriggerCOVToWriterClone(producerModel, &syncCOV)
+	if err != nil {
+		return err
+	}
+
+	data, _ := json.Marshal(point.Priority)
 	ph := new(model.ProducerHistory)
 	ph.ProducerUUID = uuid
-	ph.DataStore = writeData
-	ph.Timestamp = time.Now().UTC()
-	_, err := d.CreateProducerHistory(ph)
+	ph.DataStore = data
+	_, err = d.CreateProducerHistory(ph)
 	if err != nil {
-		return nil, err
+		log.Errorf("producer: issue on write history ProducerWriteHist: %v\n", err)
+		return errors.New("issue on write history for point")
 	}
-	return ph, nil
+	return nil
 }
 
-// ProducerWrite  update it
-func (d *GormDatabase) ProducerWrite(thingType string, payload interface{}) (string, error) {
-	var producerModel model.Producer
-	p, err := eventbus.DecodeBody(thingType, payload)
+func (d *GormDatabase) ProducersScheduleWrite(uuid string, body *model.ScheduleData) error {
+	producerModelBody := new(model.Producer)
+	producerModelBody.CurrentWriterUUID = uuid // TODO: check current_writer_uuid is needed or not
+	producers, _ := d.GetProducers(api.Args{ProducerThingUUID: &uuid})
+	for _, producer := range producers {
+		err := d.producerScheduleWrite(producer.UUID, body, producerModelBody)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (d *GormDatabase) producerScheduleWrite(uuid string, scheduleData *model.ScheduleData, producerModelBody *model.Producer) error {
+	producerModel, err := d.UpdateProducer(uuid, producerModelBody) // TODO: check current_writer_uuid is needed or not
 	if err != nil {
-		return "", err
+		log.Errorf("producer: issue on update producer err: %v\n", err)
+		return errors.New("issue on update producer")
 	}
-	if thingType == model.ThingClass.Point {
-		point := p.(*model.Point)
-		producerModel.CurrentWriterUUID = point.UUID
-		pointUUID := point.UUID
-		pro, err := d.GetProducerByField("producer_thing_uuid", pointUUID)
-		if err != nil {
-			return "", errors.New("no point for this producer was not found")
-		}
-		_, err = d.UpdateProducer(pro.UUID, &producerModel)
-		if err != nil {
-			log.Errorf("producer: issue on update producer err: %v\n", err)
-			return "", errors.New("issue on update producer")
-		}
-		value := point.PresentValue
-		point.Priority.P16 = value
-		b, err := json.Marshal(point)
-		if err != nil {
-			log.Errorf("producer: on update write history for point err: %v\n", err)
-			return "", errors.New("issue on update write history for point")
-		}
-		_, err = d.ProducerWriteHist(pro.UUID, b)
-		if err != nil {
-			log.Errorf("producer: issue on write history ProducerWriteHist: %v\n", err)
-			return "", errors.New("issue on write history for point")
-		}
-		var proBody model.ProducerBody
-		proBody.StreamUUID = pro.StreamUUID
-		proBody.ProducerUUID = pro.UUID
-		proBody.ThingType = pro.ProducerThingType
-		proBody.ThingType = point.ThingType
-		proBody.Payload = point
-		err = d.producerBroadcast(proBody)
-		if err != nil {
-			return "", errors.New("issue on producer broadcast")
-		}
-		return pro.UUID, err
+
+	syncCOV := model.SyncCOV{Schedule: scheduleData}
+	err = d.TriggerCOVToWriterClone(producerModel, &syncCOV)
+	if err != nil {
+		return err
 	}
-	return "", err
+	return nil
 }
 
-// GetProducerByField returns the point for the given field ie name or nil.
-//for example get a producer by its producer_thing_uuid
-func (d *GormDatabase) GetProducerByField(field string, value string) (*model.Producer, error) {
-	var producerModel *model.Producer
-	f := fmt.Sprintf("%s = ? ", field)
-	query := d.DB.Where(f, value).First(&producerModel)
-	if query.Error != nil {
-		return nil, query.Error
+func (d *GormDatabase) TriggerCOVToWriterClone(producer *model.Producer, body *model.SyncCOV) error {
+	wcs, err := d.GetWriterClones(api.Args{ProducerUUID: utils.NewStringAddress(producer.UUID)})
+	if err != nil {
+		return errors.New("error on getting writer clones from producer_uuid")
 	}
-	return producerModel, nil
+	for _, wc := range wcs {
+		_ = d.TriggerCOVFromWriterCloneToWriter(producer, wc, body)
+	}
+	return nil
 }
 
-// UpdateProducerByField get by field and update.
-//for example update a producer by its producer_thing_uuid
-func (d *GormDatabase) UpdateProducerByField(field string, value string, body *model.Producer, updateHist bool) (*model.Producer, error) {
-	var producerModel *model.Producer
-	f := fmt.Sprintf("%s = ? ", field)
-	query := d.DB.Where(f, value).Find(&producerModel).Updates(body)
-	if query.Error != nil {
-		return nil, query.Error
+func (d *GormDatabase) TriggerCOVFromWriterCloneToWriter(producer *model.Producer, wc *model.WriterClone, body *model.SyncCOV) error {
+	stream, _ := d.GetStream(producer.StreamUUID, api.Args{WithFlowNetworks: true})
+	body.WriterUUID = wc.SourceUUID
+	for _, fn := range stream.FlowNetworks {
+		// TODO: wc.FlowFrameworkUUID == "" remove from condition; it's here coz old deployment doesn't used to have that value
+		if wc.FlowFrameworkUUID == "" || fn.UUID == wc.FlowFrameworkUUID {
+			cli := client.NewFlowClientCli(fn.FlowIP, fn.FlowPort, fn.FlowToken, fn.IsMasterSlave, fn.GlobalUUID, model.IsFNCreator(fn))
+			_ = cli.SyncCOV(body)
+		}
 	}
-	return producerModel, nil
+	return nil
 }