+++ database/syncWriter.go
@@ -9,27 +9,25 @@ import (
 )
 
 func (d *GormDatabase) SyncWriter(body *model.SyncWriter) (*model.WriterClone, error) {
-	mWriter, err := json.Marshal(body.Writer)
-	if err != nil {
-		return nil, err
-	}
 	writerClone := model.WriterClone{}
-	if err = json.Unmarshal(mWriter, &writerClone); err != nil {
-		return nil, err
-	}
-	_, err = d.GetProducer(body.ProducerUUID, api.Args{})
+	producer, err := d.GetProducer(body.ProducerUUID, api.Args{})
 	if err != nil {
 		return nil, errors.New("producer does not exist")
 	}
+	writerClone.WriterThingName = producer.ProducerThingName
+	writerClone.WriterThingUUID = producer.ProducerThingUUID
+	writerClone.WriterThingClass = producer.ProducerThingClass
+	writerClone.WriterThingType = producer.ProducerThingType
 	writerClone.ProducerUUID = body.ProducerUUID
-	writerClone.SourceUUID = body.Writer.UUID
+	writerClone.SourceUUID = body.WriterUUID
+	writerClone.FlowFrameworkUUID = body.FlowFrameworkUUID
 	var writerCloneModel []*model.WriterClone
 	if err = d.DB.Where("source_uuid = ? ", writerClone.SourceUUID).Find(&writerCloneModel).Error; err != nil {
 		return nil, err
 	}
 	if len(writerCloneModel) == 0 {
 		writerClone.UUID = utils.MakeTopicUUID(model.CommonNaming.WriterClone)
-		if err = d.DB.Create(writerClone).Error; err != nil {
+		if err = d.DB.Create(&writerClone).Error; err != nil {
 			return nil, err
 		}
 	} else {
@@ -39,4 +37,50 @@ func (d *GormDatabase) SyncWriter(body *model.SyncWriter) (*model.WriterClone, e
 		}
 	}
 	return &writerClone, nil
+}
+
+func (d *GormDatabase) SyncCOV(body *model.SyncCOV) error {
+	writer, err := d.GetWriter(body.WriterUUID)
+	if err != nil {
+		return err
+	}
+	uuid := writer.WriterThingUUID
+	if writer.WriterThingClass == model.ThingClass.Point {
+		pointModel := model.Point{
+			CommonUUID: model.CommonUUID{UUID: uuid},
+			Priority:   body.Priority,
+		}
+		_, err = d.PointWrite(uuid, &pointModel, false)
+		return err
+	} else {
+		return d.ScheduleWrite(writer.WriterThingUUID, body.Schedule)
+	}
+}
+
+func (d *GormDatabase) SyncWriterAction(body *model.SyncWriterAction) error {
+	writerClone, err := d.GetOneWriterCloneByArgs(api.Args{SourceUUID: &body.WriterUUID})
+	if err != nil {
+		return err
+	}
+	if writerClone.WriterThingClass == model.ThingClass.Point {
+		data, _ := json.Marshal(body.Priority)
+		writerCloneBody := model.WriterClone{CommonWriter: model.CommonWriter{DataStore: data}}
+		err = d.UpdateWriterClone(writerClone, &writerCloneBody)
+		if err != nil {
+			return nil
+		}
+		point := model.Point{Priority: body.Priority}
+		_, _ = d.PointWrite(writerClone.WriterThingUUID, &point, true)
+	} else if writerClone.WriterThingClass == model.ThingClass.Schedule {
+		data, _ := json.Marshal(body.Schedule)
+		writerCloneBody := model.WriterClone{CommonWriter: model.CommonWriter{DataStore: data}}
+		err = d.UpdateWriterClone(writerClone, &writerCloneBody)
+		if err != nil {
+			return nil
+		}
+		_ = d.ScheduleWrite(writerClone.WriterThingUUID, body.Schedule)
+	} else {
+		return errors.New("no match writer thing class")
+	}
+	return err
 }