+++ database/writerClone.go
@@ -1,7 +1,6 @@
 package database
 
 import (
-	"encoding/json"
 	"github.com/NubeIO/flow-framework/api"
 	"github.com/NubeIO/flow-framework/model"
 	"github.com/NubeIO/flow-framework/utils"
@@ -21,15 +20,6 @@ func (d *GormDatabase) GetWriterClones(args api.Args) ([]*model.WriterClone, err
 	return writerClones, nil
 }
 
-func (d *GormDatabase) CreateWriterClone(body *model.WriterClone) (*model.WriterClone, error) {
-	body.UUID = utils.MakeTopicUUID(model.CommonNaming.WriterClone)
-	query := d.DB.Create(body)
-	if query.Error != nil {
-		return nil, query.Error
-	}
-	return body, nil
-}
-
 func (d *GormDatabase) GetWriterClone(uuid string) (*model.WriterClone, error) {
 	var wcm *model.WriterClone
 	query := d.DB.Where("uuid = ? ", uuid).First(&wcm)
@@ -37,6 +27,24 @@ func (d *GormDatabase) GetWriterClone(uuid string) (*model.WriterClone, error) {
 		return nil, query.Error
 	}
 	return wcm, nil
+}
+
+func (d *GormDatabase) GetOneWriterCloneByArgs(args api.Args) (*model.WriterClone, error) {
+	var wcm *model.WriterClone
+	query := d.buildWriterCloneQuery(args)
+	if err := query.First(&wcm).Error; err != nil {
+		return nil, query.Error
+	}
+	return wcm, nil
+}
+
+func (d *GormDatabase) CreateWriterClone(body *model.WriterClone) (*model.WriterClone, error) {
+	body.UUID = utils.MakeTopicUUID(model.CommonNaming.WriterClone)
+	query := d.DB.Create(body)
+	if query.Error != nil {
+		return nil, query.Error
+	}
+	return body, nil
 }
 
 func (d *GormDatabase) DeleteWriterClone(uuid string) (bool, error) {
@@ -54,47 +62,6 @@ func (d *GormDatabase) DeleteWriterClone(uuid string) (bool, error) {
 
 }
 
-func (d *GormDatabase) UpdateWriterClone(uuid string, body *model.WriterClone, updateProducer bool) (*model.WriterClone, error) {
-	var wcm *model.WriterClone
-	query := d.DB.Where("uuid = ?", uuid).Find(&wcm)
-	if query.Error != nil {
-		return nil, query.Error
-	}
-	query = d.DB.Model(&wcm).Updates(body)
-	if query.Error != nil {
-		return nil, query.Error
-	}
-	if updateProducer {
-		pro := new(model.Producer)
-		proUUID := wcm.ProducerUUID
-		pro.CurrentWriterUUID = uuid
-		p, err := d.UpdateProducer(proUUID, pro)
-		if err != nil {
-			return nil, err
-		}
-		if body.DataStore != nil {
-			_, err := d.ProducerWriteHist(proUUID, wcm.DataStore)
-			if err != nil {
-				return nil, err
-			}
-		}
-		if p.ProducerThingClass == model.ThingClass.Point {
-			pnt := new(model.Point)
-			pri := new(model.Priority)
-			err := json.Unmarshal(body.DataStore, &pri)
-			if err != nil {
-				return nil, err
-			}
-			pnt.Priority = pri
-			_, err = d.UpdatePointValue(p.ProducerThingUUID, pnt, false)
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-	return wcm, nil
-}
-
 func (d *GormDatabase) DropWriterClone() (bool, error) {
 	var wcm *model.WriterClone
 	query := d.DB.Where("1 = 1").Delete(&wcm)
@@ -107,4 +74,12 @@ func (d *GormDatabase) DropWriterClone() (bool, error) {
 	} else {
 		return true, nil
 	}
+}
+
+func (d *GormDatabase) UpdateWriterClone(writerClone *model.WriterClone, body *model.WriterClone) error {
+	query := d.DB.Model(&writerClone).Updates(body)
+	if query.Error != nil {
+		return query.Error
+	}
+	return nil
 }