+++ plugin/nube/system/schedule.go
@@ -11,27 +11,16 @@ import (
 )
 
 func (i *Instance) run() {
-
-	class, err := i.db.GetLatestProducerHistoryByProducerName("HVAC")
-	if err != nil {
+	getSch, err := i.db.GetScheduleByName("HVAC")
+	if err != nil || getSch == nil {
 		log.Errorf("system-plugin-schedule: issue on GetLatestProducerHistoryByProducerName %v\n", err)
 		return
 	}
-	ScheduleJSON, err := schedule.DecodeSchedule(class.DataStore)
+	ScheduleJSON, err := schedule.DecodeSchedule(getSch.Schedule)
 	if err != nil {
 		log.Errorf("system-plugin-schedule: issue on DecodeSchedule %v\n", err)
+		return
 	}
-
-	writer, err := i.db.GetWriter(class.CurrentWriterUUID)
-	if err != nil {
-		log.Errorf("system-plugin-schedule: issue on GetWriter %v\n", err)
-	}
-	log.Printf("system-plugin-schedule: sch %v\n", class)
-	getSch, err := i.db.GetSchedule(writer.WriterThingUUID)
-	if err != nil {
-		log.Errorf("system-plugin-schedule: issue on GetSchedule %v\n", err)
-	}
-
 	scheduleNameToCheck := "ALL" //TODO: we need a way to specify the schedule name that is being checked for.
 
 	timezone := ScheduleJSON.Config.TimeZone