+++ router/router.go
@@ -208,7 +208,6 @@ func Create(db *database.GormDatabase, vInfo *model.VersionInfo, conf *config.Co
 			historyProducerRoutes.GET("/:producer_uuid/one", historyHandler.GetLatestProducerHistoryByProducerUUID)
 			historyProducerRoutes.GET("/points", historyHandler.GetProducerHistoriesPoints)
 			historyProducerRoutes.POST("", historyHandler.CreateProducerHistory)
-			historyProducerRoutes.POST("/bulk", historyHandler.CreateBulkProducerHistory)
 			historyProducerRoutes.DELETE("/all", historyHandler.DeleteAllProducerHistories)
 			historyProducerRoutes.DELETE("/:producer_uuid", historyHandler.DeleteProducerHistoriesByProducerUUID)
 		}
@@ -290,7 +289,6 @@ func Create(db *database.GormDatabase, vInfo *model.VersionInfo, conf *config.Co
 			pointRoutes.GET("/name", pointHandler.GetPointByName)
 			pointRoutes.PATCH("/name", pointHandler.PointWriteByName)
 			pointRoutes.GET("/field/:uuid", pointHandler.GetPointByField)
-			pointRoutes.PATCH("/field/:uuid", pointHandler.UpdatePointByFieldAndUnit)
 			pointRoutes.DELETE("/:uuid", pointHandler.DeletePoint)
 			pointRoutes.DELETE("/drop", pointHandler.DropPoints)
 		}
@@ -312,13 +310,13 @@ func Create(db *database.GormDatabase, vInfo *model.VersionInfo, conf *config.Co
 			producerRoutes.PATCH("/:uuid", producerHandler.UpdateProducer)
 			producerRoutes.DELETE("/:uuid", producerHandler.DeleteProducer)
 			producerRoutes.DELETE("/drop", producerHandler.DropProducers)
+			producerRoutes.GET("/one/args", producerHandler.GetOneProducerByArgs)
 
 			producerWriterCloneRoutes := producerRoutes.Group("/writer_clones")
 			{
 				producerWriterCloneRoutes.GET("", writerCloneHandler.GetWriterClones)
 				producerWriterCloneRoutes.POST("", writerCloneHandler.CreateWriterClone)
 				producerWriterCloneRoutes.GET("/:uuid", writerCloneHandler.GetWriterClone)
-				producerWriterCloneRoutes.PATCH("/:uuid", writerCloneHandler.UpdateWriterClone)
 				producerWriterCloneRoutes.DELETE("/:uuid", writerCloneHandler.DeleteWriterClone)
 				producerWriterCloneRoutes.DELETE("/drop", writerCloneHandler.DropWriterClone)
 			}
@@ -328,7 +326,6 @@ func Create(db *database.GormDatabase, vInfo *model.VersionInfo, conf *config.Co
 		{
 			consumerRoutes.GET("", consumerHandler.GetConsumers)
 			consumerRoutes.POST("", consumerHandler.CreateConsumer)
-			consumerRoutes.POST("/wizard", consumerHandler.AddConsumerWizard)
 			consumerRoutes.GET("/:uuid", consumerHandler.GetConsumer)
 			consumerRoutes.PATCH("/:uuid", consumerHandler.UpdateConsumer)
 			consumerRoutes.DELETE("/:uuid", consumerHandler.DeleteConsumer)
@@ -343,16 +340,11 @@ func Create(db *database.GormDatabase, vInfo *model.VersionInfo, conf *config.Co
 				consumerWriterRoutes.DELETE("/:uuid", writerHandler.DeleteWriter)
 				consumerWriterRoutes.DELETE("/drop", writerHandler.DropWriters)
 			}
-
 		}
 
 		//action's writers
 		apiRoutes.POST("/writers/action/:uuid", writerHandler.WriterAction)
 		apiRoutes.POST("/writers/action/bulk", writerHandler.WriterBulkAction)
-
-		//action's writers clones
-		apiRoutes.GET("/writers/clone/:uuid", writerCloneHandler.GetWriterClone)
-		apiRoutes.PATCH("/writers/clone/:uuid", writerCloneHandler.UpdateWriterClone)
 
 		jobRoutes := apiRoutes.Group("/jobs")
 		{
@@ -390,6 +382,7 @@ func Create(db *database.GormDatabase, vInfo *model.VersionInfo, conf *config.Co
 			schRoutes.GET("/:uuid", schHandler.GetSchedule)
 			schRoutes.GET("/field", schHandler.GetScheduleByField)
 			schRoutes.PATCH("/:uuid", schHandler.UpdateSchedule)
+			schRoutes.PATCH("/write/:uuid", schHandler.ScheduleWrite)
 			schRoutes.DELETE("/:uuid", schHandler.DeleteSchedule)
 			schRoutes.DELETE("/drop", schHandler.DropSchedules)
 		}
@@ -427,6 +420,8 @@ func Create(db *database.GormDatabase, vInfo *model.VersionInfo, conf *config.Co
 			syncRoutes.POST("/flow_network", syncFlowNetworkHandler.SyncFlowNetwork)
 			syncRoutes.POST("/stream", syncStreamHandler.SyncStream)
 			syncRoutes.POST("/writer", syncWriterHandler.SyncWriter)
+			syncRoutes.POST("/cov", syncWriterHandler.SyncCOV) // clone ---> source side
+			syncRoutes.POST("/writer_action", syncWriterHandler.SyncWriterAction)
 		}
 	}